separation of concerns, isolating the ui, deprecating old main
