+++ app_ui.py
@@ -1,0 +1,35 @@
+import tkinter as tk
+from tkinter import filedialog, messagebox
+
+class SteganographyApp:
+    def __init__(self, master):
+        self.master = master
+        master.title("LSB Steganography Application")
+
+        # Set up the layout here
+        tk.Label(master, text="Select Image:").grid(row=0, column=0)
+        self.entry_image = tk.Entry(master, width=40)
+        self.entry_image.grid(row=0, column=1)
+        tk.Button(master, text="Browse", command=self.select_image).grid(row=0, column=2)
+
+        # Placeholder for additional UI elements
+        # ...
+
+    def select_image(self):
+        # This function will handle the image selection
+        file_path = filedialog.askopenfilename()
+        if file_path:
+            self.entry_image.delete(0, tk.END)
+            self.entry_image.insert(0, file_path)
+
+        # Additional methods for other functionalities will go here
+        # ...
+
+# This function can be used to run this window directly for testing
+def run():
+    root = tk.Tk()
+    app = SteganographyApp(root)
+    root.mainloop()
+
+if __name__ == "__main__":
+    run()
